# packages/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Persona API",
    "summary": "Hosted service for building and running typed-memory AI personas.",
    "version": "0.8.0"
  },
  "paths": {
    "/v1/personas": {
      "post": {
        "tags": ["personas"],
        "summary": "Create Persona",
        "description": "Create a persona from YAML; populate memory stores; auto-generate an avatar.\n\nThe avatar is generated only when the builder supplied none (D-29-3); the\ngeneration is fail-soft so create never fails on an imagegen problem\n(D-29-X-fail-soft). A user-supplied ``avatar_url`` always wins (criterion 6).",
        "operationId": "create_persona_v1_personas_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonaRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["personas"],
        "summary": "List Personas",
        "description": "List the caller's personas (paginated; RLS-scoped).",
        "operationId": "list_personas_v1_personas_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonaSummary"
                  },
                  "title": "Response List Personas V1 Personas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/personas/author": {
      "post": {
        "tags": ["personas"],
        "summary": "Author Persona",
        "description": "Generate a DRAFT persona from a description for review (D-10-2).\n\nReturns the draft envelope (YAML + clarifying questions + prompt version) \u2014\nit does NOT create a persona row. The user reviews/refines, then saves via\n``POST /v1/personas``. The flat authoring credit is deducted per call (the\ncost is the frontier-model call, not a row; D-10-8).",
        "operationId": "author_persona_v1_personas_author_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorPersonaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthoringDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/personas/author/refine": {
      "post": {
        "tags": ["personas"],
        "summary": "Refine Persona",
        "description": "Refine a draft persona by answering a clarifying question (\u00a74, D-10-2).\n\nStateless: the request carries ``round`` (refinements already applied); the\nserver rejects ``round >= 3`` as the backstop on the 3-round cap (D-10-5).\nReturns the updated draft envelope; deducts the flat authoring credit.",
        "operationId": "refine_persona_v1_personas_author_refine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefinePersonaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthoringDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/personas/recommend-tools": {
      "post": {
        "tags": ["personas"],
        "summary": "Recommend Tools",
        "description": "Recommend a ranked tool subset for a persona description (spec 26 T09).\n\nAuthoring-time assist: given the natural-language description, a single\nmid-tier call (D-26-2) returns up to 10 catalog-valid tool recommendations,\nhighest-confidence first. Reuses the description-only ``AuthorPersonaRequest``\nbody. Deducts the flat authoring credit (a mid-tier LLM call).",
        "operationId": "recommend_tools_v1_personas_recommend_tools_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorPersonaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolRecommendationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/personas/recommend-capabilities": {
      "post": {
        "tags": ["personas"],
        "summary": "Recommend Capabilities",
        "description": "Recommend a unified, provider-tagged capability set (spec 27 T10).\n\nThe D-26-10 generalisation of ``/recommend-tools``: one mid-tier call ranks\nbuilt-in tools, skills, and MCP servers together (each tagged with its\nprovider), capped at the combined maximum (D-27-13). Deducts the same flat\nauthoring credit (a mid-tier LLM call).",
        "operationId": "recommend_capabilities_v1_personas_recommend_capabilities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorPersonaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolRecommendationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/personas/{persona_id}/tools": {
      "post": {
        "tags": ["personas"],
        "summary": "Grant Tool",
        "description": "Enable a tool on the persona's allow-list via runtime consent (spec 26 T11).\n\nCalled when the user accepts a runtime tool-gap offer (T10). Adds the tool to\nthe persona's ``tools`` list (persisted in the YAML column \u2014 no migration)\nand records the grant as a versioned ``persona_self`` self-fact (force +\nconfidence \u2265 0.8 + reason, D-26-X-self-facts-consent-write-contract). Returns\nthe updated persona detail. Idempotent: re-granting an already-enabled tool\nis a no-op that still returns 200.",
        "operationId": "grant_tool_v1_personas__persona_id__tools_post",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Persona Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePersonaRequest": {
        "properties": {
          "yaml": {
            "type": "string",
            "title": "Yaml"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["yaml"],
        "title": "CreatePersonaRequest",
        "description": "Create a persona from a YAML document (validated against the v1.0 schema).\n\n``avatar_url`` is an optional presentation field (not part of the YAML\nschema) \u2014 the persona-list / chat-header visual identity."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PersonaSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "name", "role", "created_at", "updated_at"],
        "title": "PersonaSummary",
        "description": "A persona in a list view (no full YAML)."
      },
...
    }
  }
}